requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
cssselect==1.2.0
distlib==0.3.6
distro==1.8.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
filelock==3.12.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
lxml==4.9.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
openai==0.27.8
outcome==1.2.0
packaging==23.1
pandas==2.0.0
pdfminer.six==20221105
pdfquery==0.4.3
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
pycparser==2.21
pydub==0.25.1
pyquery==2.0.0
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
regex==2023.6.3
requests==2.28.2
roman==4.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tabula-py==2.7.0
tiktoken==0.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.22.0
wsproto==1.2.0
yarl==1.9.2
